NAS-127778: Move method to form
